+++ lua/mini/jump.lua
@@ -52,6 +52,18 @@
 --- disabling module's functionality is left to user. See
 --- |mini.nvim-disabling-recipes| for common recipes.
 
+--- Events ~
+---
+--- To allow user customization and integration of external tools, certain |User|
+--- autocommand events are triggered under common circumstances:
+---
+--- - `MiniJumpGetTarget` - before asking user for the target. Use |MiniJump.state|
+---   for more information about the upcoming jump.
+--- - `MiniJumpStart` - after jumping has started.
+--- - `MiniJumpJump` - after performing a jump.
+--- - `MiniJumpStop` - after jumping is stopped.
+---@tag MiniJump-events
+
 ---@alias __jump_target string|nil The string to jump to.
 ---@alias __jump_backward boolean|nil Whether to jump backward.
 ---@alias __jump_till boolean|nil Whether to jump just before/after the match instead of
@@ -218,7 +230,11 @@ MiniJump.jump = function(target, backward, till, n_times)
   -- Make jump(s)
   H.cache.n_cursor_moved = 0
   local init_cursor_data = H.get_cursor_data()
+  local was_jumping = MiniJump.state.jumping
   MiniJump.state.jumping = true
+  if not was_jumping then H.trigger_event('MiniJumpStart') end
+  H.trigger_event('MiniJumpJump')
+
   for _ = 1, MiniJump.state.n_times do
     vim.fn.search(pattern, flags)
   end
@@ -249,6 +265,8 @@ MiniJump.smart_jump = function(backward, till)
   local has_changed_cursor = not vim.deep_equal(H.cache.latest_cursor, H.get_cursor_data())
   if has_changed_mode or has_changed_cursor then MiniJump.stop_jumping() end
 
+  H.update_state(nil, backward, till, vim.v.count1)
+
   -- Ask for target only when needed
   local target
   if not MiniJump.state.jumping or MiniJump.state.target == nil then
@@ -257,7 +275,7 @@ MiniJump.smart_jump = function(backward, till)
     if target == nil then return end
   end
 
-  H.update_state(target, backward, till, vim.v.count1)
+  H.update_state(target)
 
   MiniJump.jump()
 end
@@ -270,6 +288,7 @@ MiniJump.stop_jumping = function()
   H.timers.highlight:stop()
   H.timers.idle_stop:stop()
 
+  local was_jumping = MiniJump.state.jumping
   MiniJump.state.jumping = false
 
   H.cache.n_cursor_moved = 0
@@ -277,6 +296,9 @@ MiniJump.stop_jumping = function()
   H.cache.msg_shown = false
 
   H.unhighlight()
+
+  -- Trigger relevant event only if there was jumping
+  if was_jumping then H.trigger_event('MiniJumpStop') end
 end
 
 -- Helper data ================================================================
@@ -376,6 +398,8 @@ H.make_expr_jump = function(backward, till)
   return function()
     if H.is_disabled() then return '' end
 
+    H.update_state(nil, backward, till, vim.v.count1)
+
     -- Ask for `target` for non-repeating jump as this will be used only in
     -- operator-pending mode. Dot-repeat is supported via expression-mapping.
     local is_repeat_jump = backward == nil or till == nil
@@ -383,7 +407,7 @@ H.make_expr_jump = function(backward, till)
 
     -- Stop if user supplied invalid target
     if target == nil then return '<Esc>' end
-    H.update_state(target, backward, till, vim.v.count1)
+    H.update_state(target)
 
     vim.schedule(function()
       if H.cache.has_changed_cursor then return end
@@ -527,6 +551,8 @@ end
 
 H.message = function(msg) H.echo(msg, true) end
 
+H.trigger_event = function(event_name, data) vim.api.nvim_exec_autocmds('User', { pattern = event_name, data = data }) end
+
 H.update_state = function(target, backward, till, n_times)
   MiniJump.state.mode = vim.fn.mode(1)
 
@@ -546,6 +572,8 @@ H.get_target = function()
     H.echo('Enter target single character ')
     H.cache.msg_shown = true
   end, 1000)
+
+  H.trigger_event('MiniJumpGetTarget')
   local ok, char = pcall(vim.fn.getcharstr)
   needs_help_msg = false
   H.unecho()